Encapsulate insights specific entities. A handler.js bahaves like a proxy, #123
